requirements/runtime.txt
attributee
dotty_dict
einops
lap
matplotlib
mmcls>=0.16.0,<1.0.0
motmetrics
packaging
pandas<=1.3.5
pycocotools
scipy<=1.7.3
seaborn
terminaltables
tqdm
torchreid
yolov7detect
